Add detailed Chinese comments to DAO and entity classes
